+++ src/main/java/com/sjxm/springbootinit/model/dto/DevicePageQueryDTO.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 
 @Data
 public class DevicePageQueryDTO implements Serializable {
-    private Integer schoolId;
+    private Long schoolId;
 
     private Integer inUse;
